Remove rbtree
Not using a red-black tree for the buckets. It
appears that it is actually faster to use linked
list because much faster insertions and the buckets
don't get very deep.  Maybe will add this back in
the future if I need it, but for now they definitely
add a lot of complexity.